+++ packages/launcher/lib/known-browsers.ts
@@ -156,5 +156,15 @@ export const knownBrowsers: Browser[] = [
     binary: ['edge-dev', 'microsoft-edge-dev'],
     minSupportedVersion: MIN_EDGE_VERSION,
   },
+  // @ts-ignore
   DEFAULT_ELECTRON_BROWSER as Browser,
+  {
+    name: 'webkit',
+    family: 'webkit',
+    channel: 'stable',
+    displayName: 'WebKit',
+    versionRegex: /BROWSER_VERSION = \'(?<version>[^']+)\'/gm,
+    binary: ['playwright-webkit'],
+    warning: 'WebKit support is currently experimental. Some functions may not work as expected.',
+  },
 ]