+++ cli/CHANGELOG.md
@@ -14,6 +14,8 @@ _Released 11/21/2023 (PENDING)_
 - Fixed an issue where [aliasing individual requests](https://docs.cypress.io/api/commands/intercept#Aliasing-individual-requests) with `cy.intercept()` led to an error when retrieving all of the aliases with `cy.get(@alias.all)` . Addresses [#25448](https://github.com/cypress-io/cypress/issues/25448)
 - The URL of the application under test and command error "Learn more" links now open externally instead of in the Cypress-launched browser. Fixes [#24572](https://github.com/cypress-io/cypress/issues/24572).
 - Fixed issue where some URLs would timeout in pre-request correlation. Addressed in [#28354](https://github.com/cypress-io/cypress/pull/28354).
+- Correctly detect the Webkit browser when running `cypress info` . Fixed in [#?](https://github.com/cypress-io/cypress/pull/?).
+- Display the full list of Cypress supported browsers in the error message thrown if an invalid browser name is passed via the cli argument `--browser` . Fixed in [#?](https://github.com/cypress-io/cypress/pull/?).
 
 **Misc:**